feat: tmp

+++ FE_REPO/wear/src/main/java/com/kiwa/fluffit/presentation/components/CaloriesDisplay.kt
@@ -21,7 +21,7 @@ import com.kiwa.fluffit.R
 import com.kiwa.fluffit.presentation.theme.fluffitWearFontFamily
 
 @Composable
-fun CaloriesDisplay(modifier: Modifier, calories : Int) {
+fun CaloriesDisplay(modifier: Modifier, calories : Double) {
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
         modifier = modifier.padding(end = 1.dp)
